refactor: enum 위치 변경
