+++ src/main/java/com/wanted/preonboarding/ticket/domain/entity/PerformanceSeatInfo.java
@@ -1,7 +1,7 @@
 package com.wanted.preonboarding.ticket.domain.entity;
 
-import com.wanted.preonboarding.ticket.domain.code.ActiveType;
-import com.wanted.preonboarding.ticket.domain.code.converter.ActiveTypeConverter;
+import com.wanted.preonboarding.core.code.ActiveType;
+import com.wanted.preonboarding.core.code.converter.ActiveTypeConverter;
 import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.data.annotation.CreatedDate;
