+++ src/main/java/com/wanted/preonboarding/ticket/infrastructure/repository/DiscountInfoRepository.java
@@ -1,6 +1,6 @@
 package com.wanted.preonboarding.ticket.infrastructure.repository;
 
-import com.wanted.preonboarding.ticket.domain.code.ActiveType;
+import com.wanted.preonboarding.core.code.ActiveType;
 import com.wanted.preonboarding.ticket.domain.entity.DiscountInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
